refactor: scripts and flow
