+++ package.json
@@ -26,15 +26,15 @@
     "build": "yarn stylelint && yarn lint && yarn build:components",
     "build:components": "rimraf ./dist && yarn generate && yarn typecheck && vite build && rimraf ./dist/style.css && vue-tsc -p './tsconfig.build.json' --emitDeclarationOnly && tsc-alias -p './tsconfig.build.json'",
     "build:visualize": "BUILD_VISUALIZER=true rimraf ./dist && yarn generate && vite build -m production",
-    "generate": "ts-node -P './scripts/tsconfig.json' './scripts/generate-icon-components.ts' && yarn lint:fix:generated && yarn stylelint:fix:generated",
-    "update:component-list": "ts-node -P './scripts/tsconfig.json' './scripts/update-component-list.ts' && eslint './src/component-list.ts' --fix",
+    "generate": "ts-node -P './scripts/tsconfig.json' './scripts/generate.ts' && yarn stylelint:fix:generated && yarn lint:fix:generated",
+    "update:component-list": "ts-node -P './scripts/tsconfig.json' './scripts/component-list.ts' && eslint './src/component-list.ts' --fix",
     "dev": "USE_SANDBOX=true vite",
     "test": "FORCE_COLOR=1 vitest run",
     "test:open": "vitest --ui",
     "typecheck": "vue-tsc -p './tsconfig.build.json' --noEmit",
     "lint": "eslint '**/*.{js,ts,vue}'",
     "lint:fix": "eslint '**/*.{js,ts,vue}' --fix",
-    "lint:fix:generated": "eslint './src/components/*.vue' './src/generated-component-list.ts' --fix",
+    "lint:fix:generated": "eslint './src/components/*.vue' './src/*.ts' --fix",
     "stylelint": "stylelint './src/**/*.{css,scss,vue}'",
     "stylelint:fix": "stylelint './src/**/*.{css,scss,vue}' --fix",
     "stylelint:fix:generated": "stylelint './src/components/*.vue' --fix",
