+++ scripts/component-list.ts
@@ -1,0 +1,3 @@
+import { updateComponentList } from './utilities'
+
+updateComponentList()
