+++ scripts/generate.ts
@@ -1,0 +1,3 @@
+import { generate } from './utilities'
+
+generate()
