+++ scripts/utilities/index.ts
@@ -1,5 +1,6 @@
 export * from './constants'
 export * from './strings'
 export * from './component-list'
+export { default as generate } from './generate-icon-components'
 export { default as createComponentFromSvg } from './create-component'
 export { default as getAllFiles } from './get-all-files'
